requirements.txt
requests
python-telegram-bot
groq
supabase-py
stripe
